+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 # Changelog
 
 ## [1.0] - 2025-03-27
-### Added
 - First version of Standalone-Ruby released.
 
 ## [1.1] - 2025-03-28
-### Added
 - Added support for projects with visual interfaces.
+
+## [1.2] - 2025-04-02
+- Added .cmd file extension support.
+- Added options for new supports
+- Project text outputs corrected
+- Added launcher to lib folder